ci: update install-nix-action from v27 to v29
See: NixOS/nix#10919 for a description of the issue and the fix. Our GitHub workflow file is using: cachix/install-nix-action@v27 to install nix and this action predates the fix. So I think we need to update our workflow file to use cachix/install-nix-action with a newer version that has the fix integrated. The fix was merged into NixOS::master on Sep 3, 2024 and then the next release of NixOS was v2.24.6 on Sep 10, 2024 which means that that version of NixOS should have the fix. The github action cachix/install-nix-action@v29 was released on Sep 26, 2024 and uses NixOS 2.24.7 which means that if we bump the version to v29 we should have a solution.